Move aes functions to their own file
